Extract ImageRecognitionModelBase class as a base class for classification models
