+++ api/src/commonMain/kotlin/org/jetbrains/kotlinx/dl/api/core/util/prediction.kt
@@ -1,13 +1,10 @@
 /*
- * Copyright 2020 JetBrains s.r.o. and Kotlin Deep Learning project contributors. All Rights Reserved.
+ * Copyright 2020-2022 JetBrains s.r.o. and Kotlin Deep Learning project contributors. All Rights Reserved.
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
 package org.jetbrains.kotlinx.dl.api.core.util
 
-import com.beust.klaxon.JsonArray
-import com.beust.klaxon.JsonObject
-import com.beust.klaxon.Parser
 import org.jetbrains.kotlinx.dl.api.inference.InferenceModel
 import java.util.*
 
@@ -34,25 +31,4 @@ internal fun FloatArray.indexOfMaxN(n: Int): List<Int> {
     val predictionsQueue = PriorityQueue<Int>(Comparator.comparing { index -> -this[index] })
     predictionsQueue.addAll(indices)
     return predictionsQueue.take(n)
-}
-
-/** Forms mapping of class label to class name for the ImageNet dataset. */
-public fun loadImageNetClassLabels(): Map<Int, String> {
-    val pathToIndices = "/datasets/vgg/imagenet_class_index.json"
-
-    fun parse(name: String): Any? {
-        val cls = Parser::class.java
-        return cls.getResourceAsStream(name)?.let { inputStream ->
-            return Parser.default().parse(inputStream, Charsets.UTF_8)
-        }
-    }
-
-    val classIndices = parse(pathToIndices) as JsonObject
-
-    val imageNetClassIndices = mutableMapOf<Int, String>()
-
-    for (key in classIndices.keys) {
-        imageNetClassIndices[key.toInt()] = (classIndices[key] as JsonArray<*>)[1].toString()
-    }
-    return imageNetClassIndices
 }
