+++ api/src/jvmMain/kotlin/org/jetbrains/kotlinx/dl/api/inference/imagerecognition/ImageRecognitionModel.kt
@@ -5,12 +5,11 @@
 
 package org.jetbrains.kotlinx.dl.api.inference.imagerecognition
 
-import org.jetbrains.kotlinx.dl.api.core.shape.TensorShape
-import org.jetbrains.kotlinx.dl.api.core.util.loadImageNetClassLabels
-import org.jetbrains.kotlinx.dl.api.core.util.predictTopNLabels
+import com.beust.klaxon.JsonArray
+import com.beust.klaxon.JsonObject
+import com.beust.klaxon.Parser
 import org.jetbrains.kotlinx.dl.api.inference.InferenceModel
 import org.jetbrains.kotlinx.dl.api.inference.keras.loaders.ModelType
-import org.jetbrains.kotlinx.dl.dataset.DataLoader
 import org.jetbrains.kotlinx.dl.dataset.image.ImageConverter
 import org.jetbrains.kotlinx.dl.dataset.preprocessing.Operation
 import org.jetbrains.kotlinx.dl.dataset.preprocessing.call
@@ -19,6 +18,7 @@ import org.jetbrains.kotlinx.dl.dataset.preprocessor.image.InterpolationType
 import org.jetbrains.kotlinx.dl.dataset.preprocessor.image.convert
 import org.jetbrains.kotlinx.dl.dataset.preprocessor.image.resize
 import org.jetbrains.kotlinx.dl.dataset.preprocessor.image.toFloatArray
+import org.jetbrains.kotlinx.dl.dataset.shape.TensorShape
 import java.awt.image.BufferedImage
 import java.io.File
 import java.io.IOException
@@ -27,13 +27,13 @@ import java.io.IOException
  * The light-weight API for solving Image Recognition task with one of the Model Hub models trained on ImageNet dataset.
  */
 public class ImageRecognitionModel(
-    private val internalModel: InferenceModel,
+    internalModel: InferenceModel,
     private val modelType: ModelType<out InferenceModel, out InferenceModel>
-) : InferenceModel by internalModel {
+) : ImageRecognitionModelBase<BufferedImage>(internalModel) {
     /** Class labels for ImageNet dataset. */
-    public val imageNetClassLabels: Map<Int, String> = loadImageNetClassLabels()
+    override val classLabels: Map<Int, String> = loadImageNetClassLabels()
 
-    private val preprocessing: Operation<BufferedImage, Pair<FloatArray, TensorShape>>
+    override val preprocessing: Operation<BufferedImage, Pair<FloatArray, TensorShape>>
         get() {
             val (width, height) = if (modelType.channelsFirst)
                 Pair(internalModel.inputDimensions[1], internalModel.inputDimensions[2])
@@ -52,22 +52,6 @@ public class ImageRecognitionModel(
         }
 
     /**
-     * Predicts [topK] objects for the given [image].
-     * Default preprocessing [Operation] is applied to an image.
-     *
-     * @param [image] Input image.
-     * @param [topK] Number of top ranked predictions to return
-     *
-     * @see preprocessing
-     *
-     * @return The list of pairs <label, probability> sorted from the most probable to the lowest probable.
-     */
-    public fun predictTopKObjects(image: BufferedImage, topK: Int): List<Pair<String, Float>> {
-        val inputData = preprocessing.apply(image).first
-        return internalModel.predictTopNLabels(inputData, imageNetClassLabels, topK)
-    }
-
-    /**
      * Predicts [topK] objects for the given [imageFile].
      * Default preprocessing [Operation] is applied to an image.
      *
@@ -81,20 +65,6 @@ public class ImageRecognitionModel(
     @Throws(IOException::class)
     public fun predictTopKObjects(imageFile: File, topK: Int): List<Pair<String, Float>> {
         return predictTopKObjects(ImageConverter.toBufferedImage(imageFile), topK)
-    }
-
-    /**
-     * Predicts object for the given [image].
-     * Default preprocessing [Operation] is applied to an image.
-     *
-     * @param [image] Input image.
-     * @see preprocessing
-     *
-     * @return The label of the recognized object with the highest probability.
-     */
-    public fun predictObject(image: BufferedImage): String {
-        val inputData = preprocessing.apply(image).first
-        return imageNetClassLabels[internalModel.predict(inputData)]!!
     }
 
     /**
@@ -118,4 +88,25 @@ public class ImageRecognitionModel(
     ): ImageRecognitionModel {
         return ImageRecognitionModel(internalModel.copy(copiedModelName, saveOptimizerState, copyWeights), modelType)
     }
+}
+
+/** Forms mapping of class label to class name for the ImageNet dataset. */
+public fun loadImageNetClassLabels(): Map<Int, String> {
+    val pathToIndices = "/datasets/vgg/imagenet_class_index.json"
+
+    fun parse(name: String): Any? {
+        val cls = Parser::class.java
+        return cls.getResourceAsStream(name)?.let { inputStream ->
+            return Parser.default().parse(inputStream, Charsets.UTF_8)
+        }
+    }
+
+    val classIndices = parse(pathToIndices) as JsonObject
+
+    val imageNetClassIndices = mutableMapOf<Int, String>()
+
+    for (key in classIndices.keys) {
+        imageNetClassIndices[key.toInt()] = (classIndices[key] as JsonArray<*>)[1].toString()
+    }
+    return imageNetClassIndices
 }
